Move some "basic" types into Foundation from Specification
